new USB detection and smarts
The previous USB detection wasn't perfect. Instead we'll piggyback on the
LION charger's HiZ status. If it's not trying to charge the battery, then
either the battery is missing or toast (not good!) or it doesn't have
an external power source (USB power). The first case is interesting to
know but not necessary at the moment.